PythonOrg 01 Interpreter Arg Interactive SourceCode
Tab key invokes the completion function; the current local variables, and the available module names.

For dotted expressions such as string.a, will evaluate the expression up to the final '.' and then suggest completions from the attributes of the resulting object. Note that this may execute application-defined code if an object with a __getattr__() method is part of the expression.